% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_legend.R
\name{add_legend}
\alias{add_legend}
\title{Add legend to outer margins}
\usage{
add_legend(...)
}
\arguments{
\item{...}{passed to legend}
}
\value{
legend on outer margins
}
\description{
Add legend to outer margins
}
