\name{BeginNav}
\alias{BeginNav}
\title{Write the header for the HTML navigation file (htmlize).}
\description{Initiate the navigation file for an R listing in HTML format.}
\usage{BeginNav(navcon,bgcolor="#dddddd")}
\arguments{
 \item{navcon}{The connection for writing to the navigation file.}
 \item{bgcolor}{The background color for the navigation frame.}
}
\details{
 \code{BeginNav} sets up the file with the navigation
 frame information for the HTML listing for the \code{htmlize} function.
}
\value{nil}
\author{Jim Lemon}
\keyword{misc}
