% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_mode.R
\name{bin_mode}
\alias{bin_mode}
\title{Take the Mode of Data after Binning}
\usage{
bin_mode(x, breaks = "Sturges")
}
\arguments{
\item{x}{a numeric vector}

\item{breaks}{see hist}
}
\value{
scalar
}
\description{
Take the Mode of Data after Binning
}
\examples{
bin_mode(runif(100))
}
