% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolygonPlot.R
\name{CI_lu}
\alias{CI_lu}
\title{Custom confidence intervals}
\usage{
CI_lu(Pred = NULL, CI = NULL, Fit = NULL, SE = NULL,
  level = 0.95, degf = Inf)
}
\arguments{
\item{Pred}{predict(reg, se.fit = TRUE) object}

\item{CI}{Compute CI From Regression Prediction Object}

\item{Fit}{Compute Yhat From Regression Prediction Object}

\item{SE}{Compute Standard errors}

\item{level}{confidence level}

\item{deg}{degrees of freedom in t-distribution}
}
\value{
data frame of upper and lower confidence bounds
}
\description{
Custom confidence intervals
}
\details{
Should first try
   predict(reg, interval = 'confidence', level=level) 
   note deg=inf coef estimates are approx normally distributed
}
