\name{EndHTML}
\alias{EndHTML}
\title{End an HTML file (htmlize).}
\description{Append an ending to an HTML file.}
\usage{EndHTML(con,ending=NULL)}
\arguments{
 \item{con}{The connection for writing to the HTML file.}
 \item{ending}{Any "trailer" information to be added to the file
  before closing it.}
}
\details{
 \code{EndHTML} appends a brief ending to an HTML file.
}
\value{nil}
\author{Jim Lemon}
\keyword{misc}
