% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spag_loess.R
\name{loess_list}
\alias{loess_list}
\title{Run Multiple Loess for Spaghetti Plot}
\usage{
loess_list(form, splitDF, spag_idname, split = FALSE, parallel = TRUE, ...)
}
\arguments{
\item{form}{regression formula}

\item{splitDF}{dataframe to regress on}

\item{spag_idname}{split data frame by this id}

\item{split}{split up splitDF into list of dataframes?}

\item{parallel}{use multiple cores}

\item{...}{arguments passed to loess}
}
\value{
list of loess outputs
}
\description{
Run Multiple Loess for Spaghetti Plot
}
