% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableMaker.R
\name{matJ}
\alias{matJ}
\title{Convert Matrix for stargazerJ}
\usage{
matJ(Xmat, Xrows = 1:nrow(Xmat), Xcols = 1:ncol(Xmat), Xrd = 0,
  Xfm = NULL)
}
\arguments{
\item{Xmat}{a matrix of summary regression information}

\item{Xrows}{which rows to add braces to}

\item{Xcols}{which columns to add braces to}

\item{Xrd}{how many digits to round}

\item{Xfm}{formatC format}
}
\value{
datatable
}
\description{
Convert Matrix for stargazerJ
}
