% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableMaker_trim.R
\name{mfx}
\alias{mfx}
\title{Formatting}
\usage{
mfx(
  MFXall,
  trimnames = "PCrast5",
  varnames = c("HHI", "HHI^2"),
  cnames = c("ACLED", "UCDP"),
  ROOT = TRUE,
  nvar = 3,
  colrep = 2
)
}
\arguments{
\item{nvar}{number of variables of interest}

\item{colrep}{}
}
\value{
datatable
}
\description{
Formatting
}
