% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCreg_lm.R
\name{mfxall}
\alias{mfxall}
\title{Run multiple regressions on the same dataset}
\usage{
mfxall(FORMS, ..., mc.cores = as.numeric(system("nproc", intern = TRUE)))
}
\arguments{
\item{FORMS}{list of lmtest::coeftest objects}

\item{varnames}{names of variable names}

\item{stat}{which statistic to retrieve}
}
\value{
list of summary tables
}
\description{
Run multiple regressions on the same dataset
}
