% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCreg_lm.R
\name{mfxi.lm}
\alias{mfxi.lm}
\title{Run a Regression}
\usage{
mfxi.lm(formi, datai, vcv = "standard")
}
\arguments{
\item{testlist}{list of lmtest::coeftest objects}

\item{varnames}{names of variable names}

\item{stat}{which statistic to retrieve}
}
\value{
list of summary tables
}
\description{
Run a Regression
}
