% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableMaker_trim.R
\name{mfxlist2vec}
\alias{mfxlist2vec}
\title{Table Formatting}
\usage{
mfxlist2vec(trimlist, nvar = 3)
}
\arguments{
\item{trimlist}{object from mfxtrim}

\item{nvar}{number of variables of interest}
}
\value{
list of formatted summary tables
}
\description{
Table Formatting
}
