% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableMaker_trim.R
\name{mfxtrim}
\alias{mfxtrim}
\title{Table Summarizing}
\usage{
mfxtrim(testlist, rnames, stat = "Std. Error")
}
\arguments{
\item{testlist}{list of lmtest::coeftest objects}

\item{stat}{which statistic to retrieve}

\item{varnames}{names of variable names}
}
\value{
list of summary tables
}
\description{
Table Summarizing
}
