% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolygonPlot.R
\name{polygon_add}
\alias{polygon_add}
\title{Polygon Plot}
\usage{
polygon_add(X, ci_lu, col = rgb(0, 0, 0, 0.25), bcol = NA, ...)
}
\arguments{
\item{X}{X matrix from polygon_ci}

\item{ci_lu}{matrix from polygon_ci}

\item{col, bcol}{color objects}
}
\value{
plots polygon
}
\description{
Polygon Plot
}
