% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolygonPlot.R
\name{polygon_plot}
\alias{polygon_plot}
\title{Plot the polygon}
\usage{
polygon_plot(reg, xname, xlb = "X", ylb = "Y", xlm = NULL, ylm = NULL,
  ttl = NULL, off = FALSE, ...)
}
\arguments{
\item{reg}{lm object}

\item{xname}{name of xvariable of interest}

\item{level}{confidence interval range}
}
\value{
plots polygon
}
\description{
Plot the polygon
}
