% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableMaker_trim.R
\name{root}
\alias{root}
\title{Formatting Inputs to Unit Root}
\usage{
root(MFX, mfxn1 = "HHI", mfxn2 = "HHI^2", min0 = 0, max1 = 1)
}
\arguments{
\item{MFX}{matrix of coefficients}

\item{mfxn1, mfxn2}{variable names (rows)}

\item{min0}{minimum of possible values}

\item{max1}{maximum of possible values}
}
\value{
datatable
}
\description{
Formatting Inputs to Unit Root
}
\details{
each column of MFX should provide a different set of estimates
}
