% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scatboot.R
\name{scatboot}
\alias{scatboot}
\title{M out of N Bootstrapped Loess Confidence Intervals}
\usage{
scatboot(
  x,
  y,
  breps = 100,
  mfun = function(m) {     m^(0.9) },
  confidence = 0.9,
  family = "gaussian",
  degree = 2,
  span = 2/3
)
}
\arguments{
\item{x, y}{coordinates}

\item{breps}{number of bootstrap replications}

\item{mfun}{function to define m out of n subsample}

\item{confidence}{CI interval}

\item{degree, span, family}{loess parameters}
}
\value{
list of loess outputs
}
\description{
M out of N Bootstrapped Loess Confidence Intervals
}
\details{
see "http://content.csbs.utah.edu/~rogers/datanal/R/scatboot.r"
}
