% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scattergram.R
\name{scatter_gram}
\alias{scatter_gram}
\title{Scatter Plot with Histograms}
\usage{
scatter_gram(X, Y, XBINS = NULL, xbin_scale = function(x) {     x * 2 + 0.5
  }, xbks, ybks, col = rgb(0, 0, 0, 0.5), xlb = "X", ylb = "Y",
  ttl = NULL)
}
\description{
Scatter Plot with Histograms
}
