\name{skew}
\alias{skew}
\title{
 Calculate the skew of a distribution.
}
\description{
 Calculates the skew of a distribution using a simple formula.
}
\usage{
 skew(x)
}
\arguments{
 \item{x}{A vector of numeric values, supposedly drawn from a particular distribution.}
}
\details{
 \samp{skew} calculates the asymmetry of the distribution of the values.
}
\value{The value calculated.}
\author{Jim Lemon}
\seealso{\link{sd}}
\keyword{misc}
