% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaghettiPlot.R
\name{spaghetti_lines}
\alias{spaghetti_lines}
\title{Plot Spaghetti Lines}
\usage{
spaghetti_lines(YList, Xlist, plot_col = "#00000080", lwd = 0.5, ...)
}
\arguments{
\item{YList, Xlist}{output from PrettyR::spag_loess}

\item{plot_col}{col in lines()}

\item{lwd, ...}{passed to lines()}
}
\value{
nothing
}
\description{
Plot Spaghetti Lines
}
\details{
a loess line for each list element
}
