% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaghettiPlot.R
\name{spaghetti_mean}
\alias{spaghetti_mean}
\title{Plot Average of Spaghetti Lines}
\usage{
spaghetti_mean(YList, Xlist, plot_col = 1, lwd = 2, ...)
}
\arguments{
\item{YList, Xlist}{output from PrettyR::spag_loess}

\item{plot_col, lwd}{passed to lines()}

\item{...}{passed to lines()}
}
\value{

}
\description{
Plot Average of Spaghetti Lines
}
