% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaghettiPlot.R
\name{spaghetti_plot}
\alias{spaghetti_plot}
\title{Spaghetti Plot, a loess line for each list element}
\usage{
spaghetti_plot(
  LoList,
  pdfname,
  xlb = "X",
  ylb = "Y",
  ttl = NULL,
  xln = 3,
  yln = 3,
  xlm = NULL,
  ylm = NULL,
  plot_col = rgb(0, 0, 0, 0.5),
  meanline = FALSE,
  ...
)
}
\arguments{
\item{LoList}{output from PrettyR::spag_loess}

\item{pdfname}{full name of pdf}

\item{xlb, ylb, ttl}{axis names}

\item{xln, yln}{axis lines}

\item{xlm, ylm}{axis limits}

\item{plot_col}{color of spaghetti lines}

\item{meanline}{draw the mean of the spaghettis}

\item{...}{passed to spaghetti_lines}
}
\value{

}
\description{
Spaghetti Plot, a loess line for each list element
}
\details{
seealso http://www.ats.ucla.edu/stat/r/faq/spagplot.htm
}
