% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableMaker.R
\name{stargazerJ}
\alias{stargazerJ}
\title{Table Formatting}
\usage{
stargazerJ(Xmat, ..., new.align = NULL, new.title = NULL,
  new.metatitle = NULL, new.tail = NULL, tab.out = NULL,
  new.notes = NULL, pb.tw = 1, new.omit = NULL, out = NULL,
  column.sep.width = "0pt", new.digits = 2, new.digits.extra = 5,
  new.table.placement = "H", new.other = TRUE, new.parbox = NULL)
}
\arguments{
\item{Xmat}{object from matJ}

\item{out}{NULL}

\item{new.parbox}{}
}
\value{
datatable
}
\description{
Table Formatting
}
\details{
See also matJ
}
\examples{

MATRIX <- data.frame( diag(4) )
NOTES <- paste0("NOTE", LETTERS, collapse=" ")
new.align <- "l cc @{\\\\\\\\hspace{18pt}} cc"
TITLE <- paste0("& \\\\\\\\multicolumn{2}{c}{MEOW}&\\\\\\\\multicolumn{2}{c}{MIX}\\\\\\\\\\\\\\\\",
   "& Estimate & SE & Estimate & SE")
C3     <- paste0( " Controls &", " G$\\\\&$P & & G$\\\\&$P & ", " \\\\\\\\ ")
C4     <- paste0( " F.E. &", " C$\\\\&$T & & C$\\\\&$T & ", " \\\\\\\\ ")

stargazerJ( MATRIX,
   title="Meow Mix Table",
   label="tab:meowmix",
   column.sep.width="-4pt",
   new.notes=NOTES,
   pb.tw=.7,
   new.tail=c(C3, C4),
   new.title=TITLE,
   new.align=new.align)
 
}
