% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableMaker.R
\name{stargazerJlist}
\alias{stargazerJlist}
\title{Table Formatting for Lists}
\usage{
stargazerJlist(Xlist, ..., new.align = NULL, new.title = NULL,
  new.metatitle = NULL, new.tail = NULL, tab.out = NULL,
  new.notes = NULL, pb.tw = 1, new.omit = NULL, out = NULL,
  column.sep.width = "0pt", new.digits = 2, new.digits.extra = 5,
  new.table.placement = "H", new.other = TRUE, new.parbox = NULL)
}
\description{
Table Formatting for Lists
}
\details{
different handling in part ## Xmat and Table dimensions
}
