\name{StartNav}
\alias{StartNav}
\title{Write the header for the HTML navigation file}
\description{Initiate the navigation file for an R listing in HTML format.}
\usage{StartNav(navcon,bgcolor="#dddddd")}
\arguments{
 \item{navcon}{The connection for writing to the navigation file.}
 \item{bgcolor}{The background color for the navigation frame.}
}
\details{
 \code{StartNav} sets up the file with the navigation
 frame information for the HTML listing.
}
\value{nil}
\author{Jim Lemon}
\keyword{misc}
