% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ttest.R
\name{Ttest2}
\alias{Ttest2}
\alias{t.test2}
\alias{Ttest3}
\title{Welchs t-test}
\usage{
Ttest2(m1, m2, s1, s2, n1, n2, side = 2, m0 = 0)

Ttest3(m1, m2, s1, s2, n1, n2, m0 = 0, equal.variance = FALSE)
}
\arguments{
\item{m1, m2}{the sample means}

\item{s1, s2}{the sample standard errors}

\item{n1, n2}{the sample sizes}

\item{side}{upper, lower, or 2sided}

\item{m0}{the null value for the difference in means to be tested for. Default is 0.}
}
\value{
a summary of the T-test
}
\description{
Welchs t-test

Welchs t-test, variant3
}
\examples{
n1 <- 100
x1 <- 1:n1
y1 <- 3*x1+ rnorm(n1, 0, 50)
lm1 <- summary( lm( y1~x1) )
coefs1 <- coef(lm1)[2,1:2]
n2 <- 200
x2 <- 1:n2
y2 <- 2*x2 + rnorm(n2, 0, 50)
lm2 <- summary( lm( y2~x2) )
coefs2 <- coef(lm2)[2,1:2]
Ttest2( coefs1[1], coefs2[1], coefs1[2], coefs2[2], n1, n2  )

}
