% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableMaker_trim.R
\name{UROOT}
\alias{UROOT}
\title{Formatting Inputs to Unit Root}
\usage{
UROOT(x, b1, b2, ...)
}
\arguments{
\item{x}{sequence}

\item{b1, b2}{coefficients}
}
\value{
a vector
}
\description{
Formatting Inputs to Unit Root
}
