% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StdError.R
\name{Wstat}
\alias{Wstat}
\title{Welchs T-test for a Y variable calculated from a window on each side of x0}
\usage{
Wstat(wind, dframe, x0, yvar, xvar, corr = 0, ...)
}
\arguments{
\item{wind}{fraction (percent of data to include)}

\item{dframe}{data to cut}

\item{x0}{}

\item{xvar}{}
}
\value{
list of loess outputs
}
\description{
Welchs T-test for a Y variable calculated from a window on each side of x0
}
