% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolygonPlot.R
\name{CI_lu}
\alias{CI_lu}
\title{Custom confidence intervals}
\usage{
CI_lu(Pred = NULL, CI = NULL, Fit = NULL, SE = NULL, level = 0.95,
  degf = Inf)
}
\arguments{
\item{deg}{}
}
\value{
matrix
}
\description{
Custom confidence intervals
}
\details{
Should first try
   predict(reg, interval = 'confidence', level=level)
}
