% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolygonPlot.R
\name{polygon_ci}
\alias{polygon_ci}
\title{Polygon Plot}
\usage{
polygon_ci(reg, xname, level, ...)
}
\arguments{
\item{level}{}
}
\value{
list with matrix of lower,upper confidence intervals and X variable
}
\description{
Polygon Plot
}
