% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scatboot.R
\name{scatboot}
\alias{scatboot}
\title{Bootstrapped Loess Confidence Intervals}
\usage{
scatboot(x, y, nreps = 100, confidence = 0.9, degree = 2, span = 2/3,
  family = "gaussian")
}
\arguments{
\item{x, y}{coordinates}

\item{nreps}{number of bootstrap replications}

\item{confidence}{CI interval}

\item{degree, span, family}{loess parameters}
}
\value{
list of loess outputs
}
\description{
Bootstrapped Loess Confidence Intervals
}
\details{
see "http://content.csbs.utah.edu/~rogers/datanal/R/scatboot.r"
}
