% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scattergram.R
\name{scatter_gram}
\alias{scatter_gram}
\title{Scatter Plot with Histograms}
\usage{
scatter_gram(X, Y, XBINS = NULL, xbin_scale = function(x) {     x * 2 + 0.5
  }, xbks = "Sturges", ybks = "Sturges", col = rgb(0, 0, 0, 0.5),
  xlb = "X", ylb = "Y", xrange = NA, yrange = NA, ttl = NULL)
}
\arguments{
\item{X}{vector of values}

\item{Y}{vector of values}

\item{XBINS}{bin the X,Y for plotting}

\item{xbin_scale}{how to scale XBINS}

\item{xbks, ybks}{how to make histograms}

\item{col}{color of the plot}

\item{xlb, ylb}{axis lables}

\item{ttl}{plot title}
}
\value{
plots polygon
}
\description{
Scatter Plot with Histograms
}
\examples{
scatter_gram(1:100, runif(100))
 
}
