\name{StartNav}
\alias{StartNav}
\title{Write the header for the HTML navigation file (R2html).}
\description{Initiate the navigation file for an R listing in HTML format.}
\usage{StartNav(navcon,title="R listing",bgcolor="#dddddd")}
\arguments{
 \item{navcon}{The connection for writing to the navigation file.}
 \item{title}{Title for the navigation window.}
 \item{bgcolor}{The background color for the navigation frame.}
}
\details{
 \samp{StartNav} sets up the file with the navigation
 frame information for the HTML listing for \samp{R2html}.
}
\value{nil}
\author{Philippe Grosjean}
\keyword{misc}
